misc sysadmin stuff

+++ sysadmin/deploy_production_application_server.sh
@@ -1,13 +1,13 @@
 #!/usr/bin/env bash
 
-# last updated: 2020 apr 11 for Fedora 31
+# last updated: 2022 aug 27 for Fedora 36
 
 # TODO: redis unix socket
 
 set -ex
 
-POSTGRESQL_VERSION=13.5
-REDIS_VERSION=6.2.6
+POSTGRESQL_VERSION=14.5
+REDIS_VERSION=7.0
 
 GAMOCOSM_NETWORK=gamocosm-network
 GAMOCOSM_IMAGE=gamocosm-image
@@ -17,19 +17,17 @@ SWAP_SIZE=1g
 
 function release {
 	read -p "Hit enter to continue (exit to return to script)... "
-	bash -l
+	bash
 }
-
-cd ~
-
-# make sure cron is installed
-crontab -V
 
 echo 'Please enter a new SSH port (will be changed at the very end):'
 read SSH_PORT
 
 echo 'Please enter the directory (in the home folder) containing restore files (empty for new setup):'
 read -e RESTORE_DIR
+
+echo 'Please enter the timezone file:'
+read -e TIMEZONE
 
 SSH_PRIVATE_KEY="$(realpath "$RESTORE_DIR/id_ed25519")"
 SSH_PUBLIC_KEY="$SSH_PRIVATE_KEY.pub"
@@ -67,39 +65,39 @@ if [ ! -z "$RESTORE_DIR" ]; then
 	fi
 fi
 
-# timezone
-unlink /etc/localtime
-ln -s /usr/share/zoneinfo/America/New_York /etc/localtime
+if [ -z "$TIMEZONE" ]; then
+	echo 'Timezone file not found.'
+	exit 1
+fi
 
+echo 'Updating the system...'
 #dnf -y upgrade > >(tee dnf-upgrade.stdout.log) 2> >(tee dnf-upgrade.stderr.log)
 
-# see https://wiki.archlinux.org/index.php/Swap
-touch "$SWAP"
-chattr +C "$SWAP"
-fallocate -l "$SWAP_SIZE" "$SWAP"
-chmod 600 "$SWAP"
-mkswap "$SWAP"
-echo "$SWAP none swap defaults 0 0" >> /etc/fstab
-swapon --all
+echo 'Setting the timezone...'
+unlink /etc/localtime
+ln -s "$TIMEZONE" /etc/localtime
 
-# basic tools
+echo 'Installing basic tools...'
 dnf -y install vim tmux git htop
-# services
+echo 'Installing services...'
 dnf -y install firewalld
-# containers
+echo 'Installing podman for containers...'
 dnf -y install podman
-# nginx
+echo 'Installing nginx and related tools...'
 dnf -y install nginx certbot certbot-nginx
-# for semanage and audit2allow
-dnf -y install policycoreutils-python-utils
+echo 'Installing miscellaneous tools (semanage, audit2allow, pg_dump)...'
+dnf -y install policycoreutils-python-utils postgresql
 
+echo 'Enabling firewall...'
 systemctl enable --now firewalld
 
+echo 'Setting up dotfiles...'
 git clone https://github.com/Raekye/dotfiles.git ~/dotfiles
-ln -s dotfiles/vim ~/.vim
-ln -s dotfiles/tmux ~/.tmux.conf
-./.vim/setup.sh
+ln -s ~/dotfiles/vim ~/.vim
+ln -s ~dotfiles/tmux ~/.tmux.conf
+~/.vim/setup.sh
 
+echo 'Creating/copying ssh keys...'
 if [ -z "$RESTORE_DIR" ]; then
 	ssh-keygen -t ed25519 -f ~/.ssh/id_ed25519
 else
@@ -107,6 +105,7 @@ else
 	cp "$SSH_PUBLIC_KEY" ~/.ssh/id_ed25519.pub
 fi
 
+echo 'Cloning Gamocosm repository...'
 git clone https://github.com/Gamocosm/Gamocosm.git gamocosm
 
 pushd gamocosm
@@ -116,22 +115,27 @@ cp ~/.ssh/id_ed25519.pub id_gamocosm.pub
 
 cp sysadmin/nginx.conf /etc/nginx/conf.d/gamocosm.conf
 cp sysadmin/podman@.service /etc/systemd/system/podman@.service
+cp sysadmin/daily.service /etc/systemd/system/gamocosm-daily.service
+cp sysadmin/daily.timer /etc/systemd/system/gamocosm-daily.timer
 
 if [ -z "$RESTORE_DIR" ]; then
 	cp template.env gamocosm.env
+	echo "Please update 'gamocosm.env'."
+	release
 else
 	cp "$ENV_SCRIPT" gamocosm.env
 fi
-echo 'Please update gamocosm.env'
-release
 
 source gamocosm.env
 
+mkdir backups
+
+echo 'Creating containers...'
 podman network create "$GAMOCOSM_NETWORK"
 
 podman build --tag localhost/gamocosm --file podman/Containerfile .
 
-podman create --name "$DATABASE_HOST" --network "$GAMOCOSM_NETWORK" --env "POSTGRES_USER=$DATABASE_USER" --env "POSTGRES_PASSWORD=$DATABASE_PASSWORD" "docker.io/postgres:$POSTGRESQL_VERSION"
+podman create --name "$DATABASE_HOST" --network "$GAMOCOSM_NETWORK" --publish 127.0.0.1:5433:5432 --env "POSTGRES_USER=$DATABASE_USER" --env "POSTGRES_PASSWORD=$DATABASE_PASSWORD" "docker.io/postgres:$POSTGRESQL_VERSION"
 
 podman create --name "$SIDEKIQ_REDIS_HOST" --network "$GAMOCOSM_NETWORK" "docker.io/redis:$REDIS_VERSION"
 
@@ -154,32 +158,40 @@ podman create --name gamocosm-sidekiq --network "$GAMOCOSM_NETWORK" --env-file g
 
 systemctl enable --now "podman@gamocosm-puma"
 systemctl enable --now "podman@gamocosm-sidekiq"
-systemctl enable --now nginx
+systemctl enable --now "gamocosm-daily.timer"
 
-echo '0 6 * * * bash ~gamocosm/gamocosm/sysadmin/cron.sh' | crontab -
-
-popd
-
-pushd /etc/nginx
-cp nginx.conf nginx.conf.vanilla
-echo 'Update default server block in nginx.conf'
-release
 popd
 
 mkdir /usr/share/gamocosm
 podman exec --env-file gamocosm.env gamocosm-puma rails assets:precompile
-rm -rf /usr/share/gamocosm/public
-podman cp gamocosm-puma:/gamocosm/public/ /usr/share/gamocosm
+podman cp gamocosm-puma:/gamocosm/public /usr/share/gamocosm/
 
 firewall-cmd --add-service=http
 firewall-cmd --add-service=http --permanent
 firewall-cmd --add-service=https
 firewall-cmd --add-service=https --permanent
 
+pushd /etc/nginx
+
+cp nginx.conf nginx.conf.vanilla
+echo "Update default server block in 'nginx.conf'."
+release
+
+systemctl enable --now nginx
+
+popd
+
 echo 'Setup letsencrypt/certbot.'
 release
 
-sed -i "/^#Port 22/a Port $SSH_PORT" /etc/ssh/sshd_config
+echo 'Changing ssh port...'
+
+<< EOF cat > /etc/ssh/sshd_config.d/01-gamocosm.conf
+Port $SSH_PORT
+PasswordAuthentication no
+PermitRootLogin no
+EOF
+
 semanage port -a -t ssh_port_t -p tcp "$SSH_PORT"
 firewall-cmd "--add-port=$SSH_PORT/tcp"
 firewall-cmd "--add-port=$SSH_PORT/tcp" --permanent
